requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
google-generativeai>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0